refactor(consumer): inline RebalanceListener as anon class in KafkaOffsetManager (closes #55)
The standalone RebalanceListener record was a 5-param record carrying
state that all came from KafkaOffsetManager. Two types, one parallel
constructor, and a public RebalanceListener.safeFirst() call that
leaked the manager's internals across the boundary.

Replace with an anonymous ConsumerRebalanceListener inside
KafkaOffsetManager.createRebalanceListener(). The anon class closes
over the enclosing manager's fields directly — no constructor, no
parallel state. drainCommandQueueForRevokedPartitions becomes a
private method on the manager.

Public surface unchanged: callers still get a ConsumerRebalanceListener
from OffsetManager.createRebalanceListener().

## Test churn

Deleted RebalanceListenerTest.java (it constructed the record directly
with mock state — that seam doesn't exist anymore). Five of its
11 scenarios moved into KafkaOffsetManagerTest.RebalanceTests, now
exercised through the manager's public API:

- shouldHandlePartialRevocation
- shouldDrainCommandQueueForRevokedPartitions
- shouldPreserveNonCommitCommandsDuringRevoke
- shouldRemoveCommitCommandsThatOnlyTargetRevokedPartitions
- shouldBeNoOpAfterStop

The dropped scenarios (commit-exception swallowing, empty-queue no-op,
assigned-clears-state) are either covered by other RebalanceTests
cases or low-value defensive paths.
